[BE][mod] todo 상품에 대한 리뷰 조회 페이지네이션 기능 ,별점 타입 변경

+++ src/main/java/com/doubleowner/revibe/domain/review/dto/ReviewResponseDto.java
@@ -19,7 +19,7 @@ public class ReviewResponseDto {
 
     private String content;
 
-    private Double starRate;
+    private Integer starRate;
 
     private String image;
